requirements.txt
wget
nemo_toolkit[asr]==1.20.0
transformers>=4.26.1
faster-whisper==0.9.0
git+https://github.com/m-bain/whisperX.git@49e0130e4e0c0d99d60715d76e65a71826a97109
# git+https://github.com/facebookresearch/demucs#egg=demucs
deepmultilingualpunctuation
smart_open
fastapi
uvicorn
boto3
pydantic